Transfere migrações de Bancada e CargoBancada
Fix #1798
