+++ sapl/rules/map_rules.py
@@ -280,11 +280,10 @@ rules_group_geral = {
         (parlamentares.ComposicaoMesa, __base__, __perms_publicas__),
         (parlamentares.Frente, __base__, __perms_publicas__),
         (parlamentares.Votante, __base__, __perms_publicas__),
+        (parlamentares.Bancada, __base__, __perms_publicas__),
+        (parlamentares.CargoBancada, __base__, __perms_publicas__),
         (parlamentares.Bloco, __base__, __perms_publicas__),
 
-
-        (sessao.CargoBancada, __base__, __perms_publicas__),
-        (sessao.Bancada, __base__, __perms_publicas__),
         (sessao.TipoSessaoPlenaria, __base__, __perms_publicas__),
         (sessao.TipoResultadoVotacao, __base__, __perms_publicas__),
         (sessao.TipoExpediente, __base__, __perms_publicas__),
