+++ sapl/sessao/forms.py
@@ -23,7 +23,7 @@ from sapl.utils import (RANGE_DIAS_MES, RANGE_MESES,
                         autor_modal, timezone, choice_anos_com_sessaoplenaria,
                         FileFieldCheckMixin)
 
-from .models import (Bancada, ExpedienteMateria, JustificativaAusencia,
+from .models import (ExpedienteMateria, JustificativaAusencia,
                      Orador, OradorExpediente, OrdemDia, PresencaOrdemDia, SessaoPlenaria,
                      SessaoPlenariaPresenca, TipoResultadoVotacao,
                      OcorrenciaSessao, RetiradaPauta, TipoRetiradaPauta, OradorOrdemDia, ORDENACAO_RESUMO,
@@ -257,59 +257,6 @@ class RetiradaPautaForm(ModelForm):
             expediente.save()
         retirada.save()
         return retirada
-
-
-class BancadaForm(ModelForm):
-
-    class Meta:
-        model = Bancada
-        fields = ['legislatura', 'nome', 'partido', 'data_criacao',
-                  'data_extincao', 'descricao']
-
-    def clean(self):
-        super(BancadaForm, self).clean()
-
-        if not self.is_valid():
-            return self.cleaned_data
-
-        data = self.cleaned_data
-
-        legislatura = data['legislatura']
-
-        data_criacao = data['data_criacao']
-        if data_criacao:
-            if (data_criacao < legislatura.data_inicio or
-                    data_criacao > legislatura.data_fim):
-                raise ValidationError(_("Data de criação da bancada fora do intervalo"
-                                        " de legislatura informada"))
-
-        data_extincao = data['data_extincao']
-        if data_extincao:
-            if (data_extincao < legislatura.data_inicio or
-                    data_extincao > legislatura.data_fim):
-                raise ValidationError(_("Data fim da bancada fora do intervalo de"
-                                        " legislatura informada"))
-
-        if self.cleaned_data['data_extincao']:
-            if (self.cleaned_data['data_extincao'] <
-                    self.cleaned_data['data_criacao']):
-                msg = _('Data de extinção não pode ser menor que a de criação')
-                raise ValidationError(msg)
-        return self.cleaned_data
-
-    @transaction.atomic
-    def save(self, commit=True):
-        bancada = super(BancadaForm, self).save(commit)
-        content_type = ContentType.objects.get_for_model(Bancada)
-        object_id = bancada.pk
-        tipo = TipoAutor.objects.get(content_type=content_type)
-        Autor.objects.create(
-            content_type=content_type,
-            object_id=object_id,
-            tipo=tipo,
-            nome=bancada.nome
-        )
-        return bancada
 
 
 class ExpedienteMateriaForm(ModelForm):
