+++ sapl/sessao/urls.py
@@ -1,9 +1,8 @@
 from django.conf.urls import include, url
 
 from sapl.sessao.views import (AdicionarVariasMateriasExpediente,
-                               AdicionarVariasMateriasOrdemDia, BancadaCrud,
-                               CargoBancadaCrud, ExpedienteMateriaCrud,
-                               ExpedienteView, JustificativaAusenciaCrud,
+                               AdicionarVariasMateriasOrdemDia,
+                               ExpedienteMateriaCrud, ExpedienteView, JustificativaAusenciaCrud,
                                OcorrenciaSessaoView, MateriaOrdemDiaCrud, OradorOrdemDiaCrud,
                                MesaView, OradorCrud,
                                OradorExpedienteCrud, PainelView,
@@ -99,10 +98,6 @@ urlpatterns = [
         include(TipoJustificativaCrud.get_urls())),
     url(r'^sistema/sessao-plenaria/tipo-retirada-pauta/',
         include(TipoRetiradaPautaCrud.get_urls())),
-    url(r'^sistema/bancada/',
-        include(BancadaCrud.get_urls())),
-    url(r'^sistema/cargo-bancada/',
-        include(CargoBancadaCrud.get_urls())),
     url(r'^sistema/resumo-ordenacao/',
         resumo_ordenacao,
         name='resumo_ordenacao'),
