+++ sapl/templates/sessao/layouts.yaml
@@ -82,17 +82,6 @@ OrdemDiaDetail:
   - tipo_votacao
   - observacao
 
-Bancada:
-  {% trans 'Bancada' %}:
-  - legislatura
-  - nome partido
-  - data_criacao data_extincao
-  - descricao
-
-CargoBancada:
-  {% trans 'Cargo de Bancada' %}:
-  - nome_cargo:8 cargo_unico
-
 TipoJustificativa:
   {% trans 'Tipo de Justificativa' %}:
   - descricao
