+++ sapl/templates/sistema.html
@@ -50,8 +50,8 @@
 
 	<h2>Módulo Bancadas Parlamentares</h2>
 	<div class="row">
-	<div class="col-md-6"><a href="{% url 'sapl.sessao:bancada_list' %}" class="btn btn-link">Bancadas Parlamentares</a></div>
-	<div class="col-md-6"><a href="{% url 'sapl.sessao:cargobancada_list' %}" class="btn btn-link">Cargo de Bancada Parlamentar</a></div>
+	<div class="col-md-6"><a href="{% url 'sapl.parlamentares:bancada_list' %}" class="btn btn-link">Bancadas Parlamentares</a></div>
+	<div class="col-md-6"><a href="{% url 'sapl.parlamentares:cargobancada_list' %}" class="btn btn-link">Cargo de Bancada Parlamentar</a></div>
 	<div class="col-md-6"><a href="{% url 'sapl.parlamentares:frente_list' %}" class="btn btn-link">Frente Parlamentar</a></div>
 	<div class="col-md-6"><a href="{% url 'sapl.parlamentares:bloco_list' %}" class="btn btn-link">Bloco Parlamentar</a></div>
 	</div>
